Version bump to 0.3.0. Trigger parsing completed. Can parse full Amble triggers.amble with no errors.
An issue remains if the "in rooms" condition is used on the left side of
a condition list, as the parser may be confused at where the list ends
and the next condition to the right begins. Workaround: use sets intead
of inline room lists, or use "in rooms" as last condition in list.